app/services/qr_service.py
import uuid
import qrcode
import io
import base64
from PIL import Image


def generate_gym_qr_token() -> str:
    """Generate a unique QR token for a gym."""
    return str(uuid.uuid4()).replace("-", "")


def generate_qr_image_base64(data: str) -> str:
    """Generate QR code image as base64 string."""
    qr = qrcode.QRCode(
        version=1,
        error_correction=qrcode.constants.ERROR_CORRECT_H,
        box_size=10,
        border=4,
    )
    qr.add_data(data)
    qr.make(fit=True)

    img = qr.make_image(fill_color="black", back_color="white")

    buffer = io.BytesIO()
    img.save(buffer, format="PNG")
    buffer.seek(0)

    return base64.b64encode(buffer.read()).decode("utf-8")


def generate_qr_image_bytes(data: str) -> bytes:
    """Generate QR code image as raw bytes (for file download)."""
    qr = qrcode.QRCode(
        version=1,
        error_correction=qrcode.constants.ERROR_CORRECT_H,
        box_size=10,
        border=4,
    )
    qr.add_data(data)
    qr.make(fit=True)

    img = qr.make_image(fill_color="black", back_color="white")
    buffer = io.BytesIO()
    img.save(buffer, format="PNG")
    buffer.seek(0)
    return buffer.read()
