requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
qrcode[pil]==7.4.2
pillow==10.3.0
apscheduler==3.10.4
redis==5.0.4
python-dotenv==1.0.1
firebase-admin==6.5.0
reportlab==4.2.0
